Feature: BLE 133 error handling (#178)
* upgrade pubspec.lock files + add cancel stream sub

* fix analyzer and bump version

* added completer so connect method is awaitable and completes on connected notification only

* fix getGroupElementIds null handling

* migrate to flutter_lints (except example folder for now)

* [BleManager] enhance connect method with error and timeout handling

* implement toString for BleManagerCallbacksError

* [BleScanner] handle timeout for searchForSpecificUID method

* remove redundant assignment of isProvisioningCompleted

* rename searchForSpecificUID method and its forProxy param + update doc

* expose error code for provisioning process

+++ lib/src/utils/provisioning.dart
@@ -96,33 +96,36 @@ Future<ProvisionedMeshNode> _provisioning(
       var scanTries = 0;
       while (device == null && scanTries < 6) {
         scanTries++;
-        print('attempt #$scanTries to scan for ${deviceToProvision.id}');
+        debugPrint('attempt #$scanTries to scan for ${deviceToProvision.id}');
         final scanResults = await bleScanner.provisionedNodesInRange(
-            timeoutDuration: Duration(seconds: 5)); //increase in time reduces 'Undocumented scan throttle' error
+            timeoutDuration: const Duration(seconds: 5)); //increase in time reduces 'Undocumented scan throttle' error
         try {
           device = scanResults.firstWhere((device) => device.id == deviceToProvision.id);
         } on StateError catch (e) {
-          print('not found in scan results\n$e');
+          debugPrint('not found in scan results\n$e');
         }
         if (device != null) break;
-        await Future.delayed(Duration(milliseconds: 1500));
+        await Future.delayed(const Duration(milliseconds: 1500));
       }
       if (device == null) {
-        completer.completeError(NrfMeshProvisioningException('Didn\'t find module'));
+        completer.completeError(NrfMeshProvisioningException(ProvisioningFailureCode.notFound, 'Didn\'t find module'));
       }
       events?._provisioningReconnectController.add(null);
       try {
         await bleMeshManager.connect(device!);
         provisionedMeshNode = ProvisionedMeshNode(event.meshNode!.uuid);
       } catch (e) {
-        completer.completeError(NrfMeshProvisioningException('Error in connection during provisioning process'));
+        completer.completeError(NrfMeshProvisioningException(
+            ProvisioningFailureCode.reconnection, 'Error in connection during provisioning process'));
       }
     } catch (e) {
-      completer.completeError(NrfMeshProvisioningException('error during provisioning completed listener'));
+      completer.completeError(NrfMeshProvisioningException(
+          ProvisioningFailureCode.provisioningCompleted, 'error during provisioning completed listener'));
     }
   });
   onProvisioningFailedSubscription = meshManagerApi.onProvisioningFailed.listen((event) async {
-    completer.completeError(NrfMeshProvisioningException('Failed to provision device ${deviceToProvision.id}'));
+    completer.completeError(NrfMeshProvisioningException(
+        ProvisioningFailureCode.mesh, 'Failed to provision device ${deviceToProvision.id}'));
   });
   onProvisioningStateChangedSubscription = meshManagerApi.onProvisioningStateChanged.listen((event) async {
     if (event.state == 'PROVISIONING_CAPABILITIES') {
@@ -131,8 +134,8 @@ Future<ProvisionedMeshNode> _provisioning(
           UnprovisionedMeshNode(event.meshNode!.uuid, event.meshNode!.provisionerPublicKeyXY!);
       final elementSize = await unprovisionedMeshNode.getNumberOfElements();
       if (elementSize == 0) {
-        completer
-            .completeError(NrfMeshProvisioningException('Provisioning is failed. Module does not have any elements.'));
+        completer.completeError(NrfMeshProvisioningException(
+            ProvisioningFailureCode.nodeComposition, 'Provisioning is failed. Module does not have any elements.'));
         return;
       }
       if (Platform.isAndroid) {
@@ -146,8 +149,10 @@ Future<ProvisionedMeshNode> _provisioning(
             assigned = true;
           } catch (e) {
             unicast += 1;
+            debugPrint('[Provisioning] error, incrementing unicast to $unicast');
           }
         }
+        debugPrint('[Provisioning] successfully assigned $unicast to node !');
       }
       events?._provisioningController.add(null);
       await meshManagerApi.provisioning(unprovisionedMeshNode);
@@ -200,8 +205,8 @@ Future<ProvisionedMeshNode> _provisioning(
     await completer.future;
     await meshManagerApi.cleanProvisioningData();
     await bleMeshManager.refreshDeviceCache();
-    //Added a 1,5 seconds to wait for a refreshDeviceCache.
-    await Future.delayed(Duration(milliseconds: 1500));
+    // Added a 1 second to wait for a refreshDeviceCache.
+    await Future.delayed(const Duration(seconds: 1));
     await bleMeshManager.disconnect();
     cancelProvisioningCallbackSubscription(bleMeshManager);
     return provisionedMeshNode;
@@ -230,24 +235,25 @@ void cancelProvisioningCallbackSubscription(BleMeshManager bleMeshManager) {
 Future<bool> cancelProvisioning(
     MeshManagerApi meshManagerApi, BleScanner bleScanner, BleMeshManager bleMeshManager) async {
   if (Platform.isIOS || Platform.isAndroid) {
-    print('should cancel provisioning');
+    debugPrint('should cancel provisioning');
     try {
       bleScanner.dispose();
 
       final cachedProvisionedMeshNodeUuid = await meshManagerApi.cachedProvisionedMeshNodeUuid();
       if (bleMeshManager.isProvisioningCompleted && cachedProvisionedMeshNodeUuid != null) {
         final nodes = await meshManagerApi.meshNetwork!.nodes;
-        var nodeToDelete;
+        ProvisionedMeshNode? nodeToDelete;
         try {
           nodeToDelete = nodes.firstWhere((element) => element.uuid == cachedProvisionedMeshNodeUuid);
         } on StateError catch (e) {
-          print('node not found in network\n$e');
+          debugPrint('node not found in network\n$e');
         }
 
         if (nodeToDelete != null) {
-          await meshManagerApi.deprovision(nodeToDelete);
-          //TODO: check if the deprov delete the node and remove the below code
-          await meshManagerApi.meshNetwork!.deleteNode(cachedProvisionedMeshNodeUuid);
+          final status = await meshManagerApi.deprovision(nodeToDelete);
+          if (status.success == false) {
+            await meshManagerApi.meshNetwork!.deleteNode(cachedProvisionedMeshNodeUuid);
+          }
         }
       }
       await meshManagerApi.cleanProvisioningData();
@@ -256,7 +262,7 @@ Future<bool> cancelProvisioning(
       cancelProvisioningCallbackSubscription(bleMeshManager);
       return true;
     } catch (e) {
-      print('ERROR - $e');
+      debugPrint('ERROR - $e');
       return false;
     }
   } else {
@@ -273,8 +279,26 @@ Future<ConfigNodeResetStatus> deprovision(MeshManagerApi meshManagerApi, Provisi
 }
 
 class NrfMeshProvisioningException implements Exception {
+  final ProvisioningFailureCode? code;
   final String? message;
-  NrfMeshProvisioningException([this.message]) : super();
+  NrfMeshProvisioningException([this.code, this.message]) : super();
+}
+
+enum ProvisioningFailureCode {
+  /// when an error occurs in the callback of  `PROVISIONING_COMPLETE` event
+  provisioningCompleted,
+
+  /// when the node isn't found during BLE scan upon `PROVISIONING_COMPLETE` event
+  notFound,
+
+  /// when the phone could not reconnect to node upon `PROVISIONING_COMPLETE` event
+  reconnection,
+
+  /// when the found node has unexpected node composition (nb of elements)
+  nodeComposition,
+
+  /// when `PROVISIONING_FAILED` event is triggered
+  mesh,
 }
 
 class BleMeshManagerProvisioningCallbacks extends BleMeshManagerCallbacks {
